Plugin system updates.   - More refactoring.
