+++ polymorphic_auth/apps.py
@@ -2,6 +2,7 @@ from django.apps import AppConfig
 from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.db.models.signals import post_migrate
+from django.utils.module_loading import autodiscover_modules
 
 
 def create_users(sender, **kwargs):
@@ -39,3 +40,4 @@ class AppConfig(AppConfig):
 
     def ready(self):
         post_migrate.connect(create_users, sender=self)
+        autodiscover_modules('polymorphic_auth_plugins')
