+++ polymorphic_auth/plugins/__init__.py
@@ -7,4 +7,21 @@ class PolymorphicAuthChildModelPlugin(six.with_metaclass(
     """
     Mount point for ``Polymorphic Auth`` child model plugins.
     """
-    pass
+
+    @classmethod
+    def get_plugin_for_model(cls, model):
+        model = cls.resolve_class(model)
+        for plugin in cls.plugins:
+            if plugin.get_model_class() == model:
+                return plugin
+
+    @classmethod
+    def unregister(cls, model):
+        """
+        Remove all existing plugins for a particular model
+        """
+        try:
+            plugin = cls.get_plugin_for_model(model)
+            cls.plugins.remove(plugin)
+        except ValueError:
+            pass
